pyproject.toml
[project]
name = "git-version"
version = "0.0.0"
description = "Determine version string based on git commits and tags"
authors = [
  {name="Jasper Ginn", email="jasperginn@gmail.com"}
]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "gitpython>=3.1.43",
    "pydantic>=2.9.1",
    "pytest>=8.3.3",
    "typer>=0.12.5",
]

[project.scripts]
git_version = "git_version.cli:entrypoint"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pre-commit>=3.8.0",
]

[tool.black]
line-length = 88
exclude = '''
^/(
  (
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.venv
    | _build
    | build
    | dist
    | .notebooks
    | .nox
  )
)
'''

[tool.isort]
profile = "black"
extend_skip = [".tmp", ".venv"]

[tool.mypy]
exclude = "^.tmp/"
ignore_missing_imports = true
explicit_package_bases = true

[tool.bandit]
skips = ['B101']

[tool.ruff]
lint.ignore = ["E501"]
extend-exclude = [
  "__pycache__",
  "docs",
  "site",
  ".tmp"
]

[tool.pytest.ini_options]
cache_dir = "/home/vscode/workspace/.cache/pytest"
pythonpath = ["."]

[tool.git-version]
pre_release_commit_hash = true
